refactor(auth): extract app factory helpers to remove lazy imports
